+++ components/grid/custom.scss
@@ -1,8 +1,5 @@
-@import '~ag-grid/dist/styles/ag-grid';
-@import '~ag-grid/src/styles/ag-theme-classic-common';
-@import '~ag-grid/src/styles/ag-theme-classic-vars';
-
-$icons-path: '~ag-grid/src/styles/balham-icons/';
+@import '~ag-grid-community/src/styles/ag-grid.scss';
+@import '~ag-grid-community/src/styles/ag-theme-material/sass/ag-theme-material.scss';
 
 $secondary-font-family: 'Source Sans Pro';
 $secondary-font-weight: 600;
@@ -10,13 +7,14 @@ $secondary-font-size: 16px;
 $font-family: 'Source Sans Pro';
 $font-size: 16px;
 $grid-size: 8px;
-$row-height: 36px;
-$header-height: $row-height;
+$row-height: 40px;
+$header-height: 45px;
 
 // Colors
 $icon-color: #333 !default;
 $foreground-color: #3d3d3d;
 $secondary-foreground-color: $foreground-color;
+$disabled-foreground-color-opacity: 0.5;
 $disabled-foreground-color: rgba(black, $disabled-foreground-color-opacity) !default;
 $menu-option-active-color: #d5ecfc !default;
 
@@ -68,82 +66,15 @@ $value-change-delta-down-color: darkred !default;
 $value-change-value-highlight-background-color: #cec !default;
 
 $params: (
-  grid-size: $grid-size,
-  icon-size: $icon-size,
-  header-icon-size: $header-icon-size,
-  row-height: $row-height,
-  header-height: $header-height,
-  virtual-item-height: $virtual-item-height,
-  rich-select-item-height: $row-height,
-  cell-horizontal-padding: $cell-horizontal-padding,
-  toolpanel-indent-size: $toolpanel-indent-size,
-  row-group-indent-size: $row-group-indent-size,
-  icons-path: $icons-path,
-  font-family: $font-family,
-  font-size: $font-size,
-  font-weight: $font-weight,
-  secondary-font-family: $secondary-font-family,
-  secondary-font-size: $secondary-font-size,
-  secondary-font-weight: $secondary-font-weight,
-  foreground-color: $foreground-color,
-  foreground-opacity: $foreground-opacity,
-  secondary-foreground-color-opacity: $secondary-foreground-color-opacity,
-  secondary-foreground-color: $secondary-foreground-color,
-  disabled-foreground-color-opacity: $disabled-foreground-color-opacity,
-  disabled-foreground-color: $disabled-foreground-color,
-  background-color: $background-color,
-  tab-background-color: $tab-background-color,
-  odd-row-background-color: $odd-row-background-color,
-  header-cell-moving-background-color: $header-cell-moving-background-color,
-  icon-color: $icon-color,
-  border-color: $border-color,
-  button-background-color: transparent,
-  cell-data-changed-color: $cell-data-changed-color,
-  chip-background-color: $chip-background-color,
-  editor-background-color: $editor-background-color,
-  range-selection-background-color: $range-selection-background-color,
-  range-selection-highlight-color: $range-selection-highlight-color,
-  header-background-color: $header-background-color,
-  header-background-image: $header-background-image,
-  panel-background-color: $panel-background-color,
-  tool-panel-background-color: $tool-panel-background-color,
-  card-background-color: $card-background-color,
-  accent-color: $accent-color,
-  primary-color: $primary-color,
-  hover-color: $hover-color,
-  selected-color: $selected-color,
-  icon-opacity: $foreground-opacity,
-  button-text-transform: uppercase,
-  card-radius: $card-radius,
-  card-shadow: $card-shadow,
-  input-height: $grid-size * 4,
-  focused-cell-border-color: $focused-cell-border-color,
-  menu-option-active-color: $menu-option-active-color,
-  ag-range-selected-color-1: $ag-range-selected-color-1,
-  ag-range-selected-color-2: $ag-range-selected-color-2,
-  ag-range-selected-color-3: $ag-range-selected-color-3,
-  ag-range-selected-color-4: $ag-range-selected-color-4,
-  cell-highlight-border: $cell-highlight-border,
-  cell-horizontal-border: $cell-horizontal-border,
-  value-change-delta-up-color: $value-change-delta-up-color,
-  value-change-delta-down-color: $value-change-delta-down-color,
-  value-change-value-highlight-background-color: $value-change-value-highlight-background-color,
-  row-floating-background-color: $row-floating-background-color,
-  row-stub-background-color: $row-stub-background-color,
+	grid-size: $grid-size,
+	row-height: $row-height,
+	header-height: $header-height,
 );
 
 .ag-theme-faithlife {
-  @include ag-theme-classic($params);
+	@extend .ag-theme-material;
 
-  %card {
-      border: 1px solid #ccc;
-  }
-}
-
-.ag-cell {
-  border: none !important;
-}
-
-.ag-cell-right-aligned {
-  text-align: right;
+	%card {
+		border: 1px solid #ccc;
+	}
 }
