+++ components/grid/index.js
@@ -1,1 +1,4 @@
 import './custom.scss';
+export { PaginatedTable } from './paginated-table';
+export { TableHeading } from './table-helpers';
+export { SimpleTable } from './simple-table';
