+++ components/grid/simple-table.jsx
@@ -1,26 +1,19 @@
-import React, { useState, useMemo } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
+import { useTableState } from './table-helpers';
 import { BaseTable } from './base-table';
 
-export function InfiniteScrollTable({
+export function SimpleTable({
 	onRowClick,
-	isLoading,
 	children,
 	isSmallViewport,
 	data,
 	filterText,
 	sortModel,
 	updateSortModel,
+	maxRows,
 }) {
-	const [gridApi, setGridApi] = useState(null);
-	const [columnApi, setColumnApi] = useState(null);
-
-	/*const gridOptions = useMemo(
-		() => ({
-			domLayout: 'autoHeight',
-		}),
-		[],
-	);*/
+	const { gridApi, setGridApi, columnApi, setColumnApi } = useTableState();
 
 	return (
 		<BaseTable
@@ -34,6 +27,7 @@ export function InfiniteScrollTable({
 			data={data}
 			updateSortModel={updateSortModel}
 			filterText={filterText}
+			maxRowsPerPage={maxRows}
 		>
 			{children}
 		</BaseTable>
@@ -41,7 +35,7 @@ export function InfiniteScrollTable({
 }
 
 /*
-InfiniteScrollTable.propTypes = {
+SimpleTable.propTypes = {
 	onRowClick: PropTypes.func,
 	largeOnlyColumns: PropTypes.arrayOf(PropTypes.string),
 	maxRowsPerPage: PropTypes.number,
