--- components/table/index.js
@@ -1,3 +1,0 @@
-export { PaginatedTable } from './paginated-table';
-export { InfiniteScrollTable } from './infinite-scroll-table';
-export { TableHeading } from './table-helpers';
